Update Stack_ROI_Measure.ijm
This macro is under development as a tool to automatically detect and measure changes in the position and size of particles (e.g. dust and glass corrosion products) on daguerreotype plates and cover glasses. The macro could have broader art conservation applications, possible for monitoring changes before and after exhibition of other mostly 2-dimensional media.